requirements.txt
requests>=2.31.0
playwright>=1.40.0
gspread>=6.0.0
python-dotenv>=1.0.0
google-api-python-client>=2.0.0
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
google-generativeai>=0.3.0
beautifulsoup4>=4.12.0
deep_translator>=1.11.0
curl_cffi>=0.5.0
openpyxl>=3.1.0
